applied mysql changes to JPA

+++ src/main/java/com/samuraiDigital/adminsystem/data/repositories/SecurityGroupRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.samuraiDigital.adminsystem.data.model.SecurityGroup;
 
-public interface SecurityGroupRepository extends CrudRepository<SecurityGroup, Long>{
+public interface SecurityGroupRepository extends CrudRepository<SecurityGroup, Integer>{
 
 	Optional<SecurityGroup> findByName(String name);
