+++ src/main/java/com/samuraiDigital/adminsystem/data/repositories/UserSecurityDetailsRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.samuraiDigital.adminsystem.data.model.UserSecurityDetails;
 
-public interface UserSecurityDetailsRepository extends CrudRepository<UserSecurityDetails, Long>{
+public interface UserSecurityDetailsRepository extends CrudRepository<UserSecurityDetails, Integer>{
 	
 	Optional<UserSecurityDetails> findByUsername(String username);
